Actually use MaxAttempts on NSQ Config instance.

+++ internal/component/input/config_nsq.go
@@ -13,7 +13,7 @@ type NSQConfig struct {
 	UserAgent       string      `json:"user_agent" yaml:"user_agent"`
 	TLS             btls.Config `json:"tls" yaml:"tls"`
 	MaxInFlight     int         `json:"max_in_flight" yaml:"max_in_flight"`
-	MaxAttempts     int         `json:"max_attempts" yaml:"max_attempts"`
+	MaxAttempts     uint16      `json:"max_attempts" yaml:"max_attempts"`
 }
 
 // NewNSQConfig creates a new NSQConfig with default values.
